Added support for fp16 and bf16
- Added fp16 to doctests
- using input and backend param dtype for pybinding
- NWE input has to be float for now

+++ test/python/src/mlp.py
@@ -15,7 +15,7 @@ class MLP(torch.nn.Module):
         activation_func="relu",
         output_activation=None,
         use_batchnorm=False,
-        dtype=torch.float,
+        dtype=torch.bfloat16,
     ):
         super().__init__()
         self.dtype = dtype
